+++ common/dataStructures/LabelFormat.hpp
@@ -17,46 +17,46 @@
 
 #pragma once
 
-#include "common/exception/InvalidArgument.hpp"
-
-#include <sstream>
 #include <iomanip>
 #include <optional>
+#include <sstream>
+#include <string>
+
+#include "common/exception/InvalidArgument.hpp"
 
 namespace cta {
 namespace common {
 namespace dataStructures {
 
-  struct Label {
+struct Label {
+  enum class Format : std::uint8_t {
+    CTA = 0x00
+  };
 
-    enum class Format : std::uint8_t {
-      CTA = 0x00
-    };
+  static Format validateFormat(const std::optional<std::uint8_t>& ouiFormat, const std::string& strContext) {
+    return validateFormat(ouiFormat.value_or(static_cast<uint8_t>(Format::CTA)), strContext);
+  }
 
-    static Format validateFormat(const std::optional<std::uint8_t>& ouiFormat, const std::string& strContext) {
-      return validateFormat(ouiFormat.value_or(0x00), strContext);
-    }
-
-    static Format validateFormat(const std::uint8_t uiFormat, const std::string& strContext) {
-      Format format = static_cast<Format>(uiFormat);
-      switch(format) {
-        case Format::CTA:
-          return format;
-        default:
-          {
-            std::ostringstream ex_str;
-            ex_str << strContext << ": Unknown label format " << std::showbase << std::internal << std::setfill('0') << std::hex << std::setw(4) << static_cast<unsigned int>(uiFormat);
-            throw cta::exception::InvalidArgument(ex_str.str());
-          }
+  static Format validateFormat(const std::uint8_t uiFormat, const std::string& strContext) {
+    Format format = static_cast<Format>(uiFormat);
+    switch (format) {
+      case Format::CTA:
+        return format;
+      default:
+      {
+        std::ostringstream ex_str;
+        ex_str << strContext << ": Unknown label format " << std::showbase << std::internal << std::setfill('0')
+                << std::hex << std::setw(4) << static_cast<unsigned int>(uiFormat);
+        throw cta::exception::InvalidArgument(ex_str.str());
       }
     }
-    // prevent the generation of default public constructor and destructor
-    protected:
-      Label() = default;
-      virtual ~Label() = default;
+  }
+  // prevent the generation of default public constructor and destructor
+protected:
+  Label() = default;
+  virtual ~Label() = default;
+};
 
- };
-
-} // namespace dataStructures
-} // namespace common
-} // namespace cta
+}  // namespace dataStructures
+}  // namespace common
+}  // namespace cta