+++ tapeserver/castor/tape/tapeserver/file/FileReader.hpp
@@ -31,10 +31,40 @@ namespace tape {
 namespace tapeFile {
 
 class ReadSession;
-class UHL1;
 
 class FileReader{
- public:
+public:
+  /**
+    * Destructor of the FileReader. It will release the lock on the read session.
+    */
+  virtual ~FileReader() throw();
+
+  /**
+    * After positioning at the beginning of a file for readings, this function
+    * allows the reader to know which block sizes to provide.
+    * @return the block size in bytes.
+    */
+  size_t getBlockSize();
+
+  /**
+    * Reads the next data block from the file. The buffer should be equal to or bigger than the
+    * block size.
+    * It has to throw EndOfFile() exception when it reache the End Of File.
+    * @param data: pointer to the data buffer
+    * @param size: size of the buffer
+    * @return The amount of data actually copied. Zero at end of file.
+    */
+  virtual size_t readNextDataBlock(void *data, const size_t size) = 0;
+
+  /**
+    * Returns the LBP access mode.
+    * @return The LBP mode.
+    */
+  std::string getLBPMode();
+
+  friend class FileReaderFactory;
+
+protected:
   /**
     * Constructor of the FileReader. It will bind itself to an existing read session
     * and position the tape right at the beginning of the file
@@ -45,51 +75,14 @@ class FileReader{
   FileReader(const std::unique_ptr<ReadSession> &rs, const cta::RetrieveJob &fileToRecall);
 
   /**
-    * Destructor of the FileReader. It will release the lock on the read session.
+    * Move throught the headers to the block to read using Fseq Delta increments.
     */
-  ~FileReader() throw();
+  virtual void positionByFseq(const cta::RetrieveJob &fileToRecall) = 0;
 
   /**
-    * After positioning at the beginning of a file for readings, this function
-    * allows the reader to know which block sizes to provide.
-    * @return the block size in bytes.
+    * Move throught the headers to the blocks to read using the ID of the block.
     */
-  size_t getBlockSize();
-
-  /**
-    * Reads data from the file. The buffer should be equal to or bigger than the 
-    * block size.
-    * @param data: pointer to the data buffer
-    * @param size: size of the buffer
-    * @return The amount of data actually copied. Zero at end of file.
-    */
-  size_t read(void *data, const size_t size);
-
-  /**
-    * Returns the LBP access mode.
-    * @return The LBP mode.
-    */
-  std::string getLBPMode();
-
- private:
-  void positionByFseq(const cta::RetrieveJob &fileToRecall);
-  void positionByBlockID(const cta::RetrieveJob &fileToRecall);
-  /**
-    * Positions the tape for reading the file. Depending on the previous activity,
-    * it is the duty of this function to determine how to best move to the next
-    * file. The positioning will then be verified (header will be read). 
-    * As usual, exception is thrown if anything goes wrong.
-    * @param fileInfo: all relevant information passed by the stager about the file.
-    */
-  void position(
-    const cta::RetrieveJob &fileToRecall);
-
-  /**
-    * Set the block size member using the info contained within the uhl1
-    * @param uhl1: the uhl1 header of the current file
-    * @param fileInfo: the Information structure of the current file
-    */
-  void setBlockSize(const UHL1 &uhl1);
+  virtual void positionByBlockID(const cta::RetrieveJob &fileToRecall) = 0;
 
   /**
     * Block size in Bytes of the current file
@@ -101,6 +94,7 @@ class FileReader{
     */
   const std::unique_ptr<ReadSession> &m_session;
 
+private:
   /**
     * What kind of command we use to position ourself on the tape (fseq or blockid)
     */
@@ -110,6 +104,15 @@ class FileReader{
     * Description of the LBP mode with which the files is read.
     */
   std::string m_LBPMode;
+
+  /**
+    * Positions the tape for reading the file. Depending on the previous activity,
+    * it is the duty of this function to determine how to best move to the next
+    * file. The positioning will then be verified (header will be read).
+    * As usual, exception is thrown if anything goes wrong.
+    * @param fileInfo: all relevant information passed by the stager about the file.
+    */
+  void position(const cta::RetrieveJob &fileToRecall);
 };
 
 }  // namespace tapeFile